requirements.txt
easydict
pyarrow
lightning
tensorboard
numpy
pandas
ipykernel
ipywidgets
scikit-learn
tqdm
albumentations
matplotlib
torcheval
tensorflow